Fix validation API to support dry run (#3455)
* Refactor unstructured list logic to make dry-run support easier to add
* Add dry run support
* Handle the multi vs list
* Properly handle concurrency
* Add changelog
* Merge branch 'master' into fix_validation_dry_run
* Merge branch 'master' into fix_validation_dry_run

+++ projects/gateway/pkg/validation/validator.go
@@ -6,7 +6,12 @@ import (
 	"sync"
 	"time"
 
+	"github.com/hashicorp/go-multierror"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+
 	gloov1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
+	skprotoutils "github.com/solo-io/solo-kit/pkg/utils/protoutils"
 
 	"github.com/avast/retry-go"
 	"github.com/solo-io/go-utils/protoutils"
@@ -36,6 +41,10 @@ var (
 	VirtualServiceDeleteErr = func(parentGateways []core.ResourceRef) error {
 		return errors.Errorf("Deletion blocked because active Gateways reference this Virtual Service. Remove refs to this virtual service from the gateways: %v, then try again", parentGateways)
 	}
+	unmarshalErrMsg     = "could not unmarshal raw object"
+	WrappedUnmarshalErr = func(err error) error {
+		return errors.Wrapf(err, unmarshalErrMsg)
+	}
 )
 
 const (
@@ -45,11 +54,12 @@ const (
 
 type Validator interface {
 	v1.ApiSyncer
-	ValidateGateway(ctx context.Context, gw *v1.Gateway) (ProxyReports, error)
-	ValidateVirtualService(ctx context.Context, vs *v1.VirtualService) (ProxyReports, error)
-	ValidateDeleteVirtualService(ctx context.Context, vs core.ResourceRef) error
-	ValidateRouteTable(ctx context.Context, rt *v1.RouteTable) (ProxyReports, error)
-	ValidateDeleteRouteTable(ctx context.Context, rt core.ResourceRef) error
+	ValidateList(ctx context.Context, ul *unstructured.UnstructuredList, dryRun bool) (ProxyReports, error)
+	ValidateGateway(ctx context.Context, gw *v1.Gateway, dryRun bool) (ProxyReports, error)
+	ValidateVirtualService(ctx context.Context, vs *v1.VirtualService, dryRun bool) (ProxyReports, error)
+	ValidateDeleteVirtualService(ctx context.Context, vs core.ResourceRef, dryRun bool) error
+	ValidateRouteTable(ctx context.Context, rt *v1.RouteTable, dryRun bool) (ProxyReports, error)
+	ValidateDeleteRouteTable(ctx context.Context, rt core.ResourceRef, dryRun bool) error
 }
 
 type validator struct {
@@ -143,16 +153,14 @@ func (v *validator) deleteFromLocalSnapshot(resource resources.Resource) {
 	}
 }
 
-func (v *validator) validateSnapshot(ctx context.Context, apply applyResource) (ProxyReports, error) {
+func (v *validator) validateSnapshot(ctx context.Context, apply applyResource, dryRun, acquireLock bool) (ProxyReports, error) {
 	if !v.ready() {
 		return nil, NotReadyErr
 	}
 
 	ctx = contextutils.WithLogger(ctx, "gateway-validator")
 
-	v.lock.RLock()
 	snap := v.latestSnapshot.Clone()
-	v.lock.RUnlock()
 
 	if v.latestSnapshotErr != nil {
 		contextutils.LoggerFrom(ctx).Errorw(InvalidSnapshotErrMessage, zap.Error(v.latestSnapshotErr))
@@ -230,15 +238,96 @@ func (v *validator) validateSnapshot(ctx context.Context, apply applyResource) (
 
 	contextutils.LoggerFrom(ctx).Debugw("Accepted %T %v", resource, ref)
 
-	// update internal snapshot to handle race where a lot of resources may be applied at once, before syncer updates
-	v.lock.Lock()
-	apply(v.latestSnapshot)
-	v.lock.Unlock()
+	if !dryRun {
+		// update internal snapshot to handle race where a lot of resources may be applied at once, before syncer updates
+		if acquireLock {
+			v.lock.Lock()
+		}
+		apply(v.latestSnapshot)
+		if acquireLock {
+			v.lock.Unlock()
+		}
+	}
 
 	return proxyReports, nil
 }
 
-func (v *validator) ValidateVirtualService(ctx context.Context, vs *v1.VirtualService) (ProxyReports, error) {
+func (v *validator) ValidateList(ctx context.Context, ul *unstructured.UnstructuredList, dryRun bool) (ProxyReports, error) {
+	var (
+		proxyReports = ProxyReports{}
+		errs         = &multierror.Error{}
+	)
+
+	v.lock.Lock()
+	defer v.lock.Unlock()
+	snap := v.latestSnapshot.Clone()
+
+	for _, item := range ul.Items {
+
+		gv, err := schema.ParseGroupVersion(item.GetAPIVersion())
+		if err != nil {
+			return ProxyReports{}, err
+		}
+
+		itemGvk := schema.GroupVersionKind{
+			Version: gv.Version,
+			Group:   gv.Group,
+			Kind:    item.GetKind(),
+		}
+
+		jsonBytes, err := item.MarshalJSON()
+		if err != nil {
+			return ProxyReports{}, err
+		}
+
+		var itemProxyReports ProxyReports
+		switch itemGvk {
+		case v1.GatewayGVK:
+			var (
+				gw v1.Gateway
+			)
+			if err := skprotoutils.UnmarshalResource(jsonBytes, &gw); err != nil {
+				return nil, WrappedUnmarshalErr(err)
+			}
+			itemProxyReports, err = v.validateGatewayInternal(ctx, &gw, false, false)
+		case v1.VirtualServiceGVK:
+			var (
+				vs v1.VirtualService
+			)
+			if err := skprotoutils.UnmarshalResource(jsonBytes, &vs); err != nil {
+				return nil, WrappedUnmarshalErr(err)
+			}
+			itemProxyReports, err = v.validateVirtualServiceInternal(ctx, &vs, false, false)
+		case v1.RouteTableGVK:
+			var (
+				rt v1.RouteTable
+			)
+			if err := skprotoutils.UnmarshalResource(jsonBytes, &rt); err != nil {
+				return nil, WrappedUnmarshalErr(err)
+			}
+			itemProxyReports, err = v.validateRouteTableInternal(ctx, &rt, false, false)
+		}
+
+		errs = multierror.Append(errs, err)
+		for proxy, report := range itemProxyReports {
+			// ok to return final proxy reports as the latest result includes latest proxy calculated
+			// for each resource, as we process incrementally, storing new state in memory as we go
+			proxyReports[proxy] = report
+		}
+	}
+
+	if dryRun {
+		v.latestSnapshot = &snap
+	}
+
+	return proxyReports, errs.ErrorOrNil()
+}
+
+func (v *validator) ValidateVirtualService(ctx context.Context, vs *v1.VirtualService, dryRun bool) (ProxyReports, error) {
+	return v.validateVirtualServiceInternal(ctx, vs, dryRun, true)
+}
+
+func (v *validator) validateVirtualServiceInternal(ctx context.Context, vs *v1.VirtualService, dryRun, acquireLock bool) (ProxyReports, error) {
 	apply := func(snap *v1.ApiSnapshot) ([]string, resources.Resource, core.ResourceRef) {
 		vsRef := vs.GetMetadata().Ref()
 
@@ -260,10 +349,10 @@ func (v *validator) ValidateVirtualService(ctx context.Context, vs *v1.VirtualSe
 		return proxiesForVirtualService(snap.Gateways, vs), vs, vsRef
 	}
 
-	return v.validateSnapshot(ctx, apply)
+	return v.validateSnapshot(ctx, apply, dryRun, acquireLock)
 }
 
-func (v *validator) ValidateDeleteVirtualService(ctx context.Context, vsRef core.ResourceRef) error {
+func (v *validator) ValidateDeleteVirtualService(ctx context.Context, vsRef core.ResourceRef, dryRun bool) error {
 	if !v.ready() {
 		return errors.Errorf("Gateway validation is yet not available. Waiting for first snapshot")
 	}
@@ -304,12 +393,17 @@ func (v *validator) ValidateDeleteVirtualService(ctx context.Context, vsRef core
 		contextutils.LoggerFrom(ctx).Debugw("Accepted deletion of Virtual Service %v", vsRef)
 	}
 
-	v.deleteFromLocalSnapshot(vs)
-
+	if !dryRun {
+		v.deleteFromLocalSnapshot(vs)
+	}
 	return nil
 }
 
-func (v *validator) ValidateRouteTable(ctx context.Context, rt *v1.RouteTable) (ProxyReports, error) {
+func (v *validator) ValidateRouteTable(ctx context.Context, rt *v1.RouteTable, dryRun bool) (ProxyReports, error) {
+	return v.validateRouteTableInternal(ctx, rt, dryRun, true)
+}
+
+func (v *validator) validateRouteTableInternal(ctx context.Context, rt *v1.RouteTable, dryRun, acquireLock bool) (ProxyReports, error) {
 	apply := func(snap *v1.ApiSnapshot) ([]string, resources.Resource, core.ResourceRef) {
 		rtRef := rt.GetMetadata().Ref()
 
@@ -333,10 +427,10 @@ func (v *validator) ValidateRouteTable(ctx context.Context, rt *v1.RouteTable) (
 		return proxiesToConsider, rt, rtRef
 	}
 
-	return v.validateSnapshot(ctx, apply)
+	return v.validateSnapshot(ctx, apply, dryRun, acquireLock)
 }
 
-func (v *validator) ValidateDeleteRouteTable(ctx context.Context, rtRef core.ResourceRef) error {
+func (v *validator) ValidateDeleteRouteTable(ctx context.Context, rtRef core.ResourceRef, dryRun bool) error {
 	if !v.ready() {
 		return errors.Errorf("Gateway validation is yet not available. Waiting for first snapshot")
 	}
@@ -377,12 +471,17 @@ func (v *validator) ValidateDeleteRouteTable(ctx context.Context, rtRef core.Res
 		contextutils.LoggerFrom(ctx).Debugw("Accepted Route Table deletion %v", rtRef)
 	}
 
-	v.deleteFromLocalSnapshot(rt)
-
+	if !dryRun {
+		v.deleteFromLocalSnapshot(rt)
+	}
 	return nil
 }
 
-func (v *validator) ValidateGateway(ctx context.Context, gw *v1.Gateway) (ProxyReports, error) {
+func (v *validator) ValidateGateway(ctx context.Context, gw *v1.Gateway, dryRun bool) (ProxyReports, error) {
+	return v.validateGatewayInternal(ctx, gw, dryRun, true)
+}
+
+func (v *validator) validateGatewayInternal(ctx context.Context, gw *v1.Gateway, dryRun, acquireLock bool) (ProxyReports, error) {
 	apply := func(snap *v1.ApiSnapshot) ([]string, resources.Resource, core.ResourceRef) {
 		gwRef := gw.GetMetadata().Ref()
 
@@ -406,7 +505,7 @@ func (v *validator) ValidateGateway(ctx context.Context, gw *v1.Gateway) (ProxyR
 		return proxiesToConsider, gw, gwRef
 	}
 
-	return v.validateSnapshot(ctx, apply)
+	return v.validateSnapshot(ctx, apply, dryRun, acquireLock)
 }
 
 func proxiesForVirtualService(gwList v1.GatewayList, vs *v1.VirtualService) []string {
